pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "contextvault"
version = "0.1.0"
description = "Local-first context management layer for AI models"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "ContextVault Contributors"}
]
keywords = ["ai", "context", "local", "privacy", "ollama", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "httpx>=0.25.2",
    "aiofiles>=23.2.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "orjson>=3.9.10",
    "cryptography>=41.0.8",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]
embeddings = [
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.2",
]

[project.urls]
Homepage = "https://github.com/contextvault/contextvault"
Documentation = "https://contextvault.readthedocs.io"
Repository = "https://github.com/contextvault/contextvault"
"Bug Tracker" = "https://github.com/contextvault/contextvault/issues"

[project.scripts]
contextvault = "contextvault.cli.main:cli"
contextible = "contextvault.cli.main:cli"

[tool.setuptools]
packages = ["contextvault", "contextvault.api", "contextvault.models", "contextvault.services", "contextvault.integrations", "contextvault.integrations.mcp", "contextvault.schemas", "contextvault.cli", "contextvault.cli.commands", "contextvault.routing", "contextvault.thinking", "contextvault.dashboard"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["contextvault"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
